Move TestExtremeCoordinatesFilter into dynamic data
